Add license headers, add sprintf to utils
